requirements.txt
pyyaml == 6.0
torch == 2.1.2
torchvision == 0.16.2
scikit-learn == 1.3.0